Ben's review

+++ core/ops.rs
@@ -80,7 +80,7 @@ where
   }
 }
 
-pub type PromiseId = u64;
+pub type PromiseId = i32;
 pub type OpAsyncFuture = OpCall<(PromiseId, OpId, OpResult)>;
 pub type OpFn = dyn Fn(Rc<RefCell<OpState>>, OpPayload) -> Op + 'static;
 pub type OpId = usize;
